refactor: Rename Objectives folder to Goals
